separating csi validation
